fix/bsc: setup script

+++ deployments/testnet/init_bsc.sh
@@ -1,3 +1,5 @@
 #!/bin/bash
 
-DIR=$DIR docker-compose run --rm bsc --datadir /home/geth init /home/geth/genesis.json
+if ! [ -e "$dir/testnet/bsc_testnet/genesis.json" ]; then
+    DIR=$DIR docker-compose run --rm bsc --datadir /home/geth init /home/geth/genesis.json
+fi
